refactor(back-end): new folder structure
