+++ server/package.json
@@ -27,6 +27,7 @@
     "bcrypt": "^5.0.0",
     "body-parser": "^1.19.0",
     "chalk": "^2.4.2",
+    "compression": "^1.7.4",
     "connect-mongo": "^3.2.0",
     "cors": "^2.8.5",
     "dotenv": "^6.2.0",
@@ -40,6 +41,7 @@
   "devDependencies": {
     "@types/bcrypt": "^3.0.0",
     "@types/body-parser": "^1.19.0",
+    "@types/compression": "^1.7.0",
     "@types/cors": "^2.8.9",
     "@types/express": "^4.17.9",
     "@types/express-session": "1.17.0",
