+++ server/src/routes/auth/auth.ts
@@ -1,0 +1,7 @@
+import app from '../../index'
+import auth from './auth.controller'
+
+app
+  .post('/register', auth.register)
+  .post('/login', auth.login)
+  .post('/logout', auth.logout)
