+++ server/src/services/db.ts
@@ -1,15 +1,16 @@
 import MongoClient, { InsertOneWriteOpResult, UpdateWriteOpResult } from 'mongodb'
 import chalk from 'chalk'
+
 import { User } from '../../../shared/interfaces/User'
 
 export let db: MongoClient.Db
 export let client: MongoClient.MongoClient
 
-export function parseId (id?: string): MongoClient.ObjectId {
+function parseId (id?: string): MongoClient.ObjectId {
   return new MongoClient.ObjectId(id)
 }
 
-export function init (database: string): Promise<void> {
+function init (database: string): Promise<void> {
   return new Promise((resolve, reject) => {
     MongoClient.connect(`mongodb://localhost:27017/${database}`, {
       useNewUrlParser: true,
@@ -27,7 +28,7 @@ export function init (database: string): Promise<void> {
   })
 }
 
-export function get (query: object): Promise<User[]> {
+function get (query: object): Promise<User[]> {
   return new Promise((resolve, reject) => {
     db.collection('users').find(query).toArray((err, result) => {
       if (err) reject(err)
@@ -36,10 +37,18 @@ export function get (query: object): Promise<User[]> {
   })
 }
 
-export function insert (data: any): Promise<InsertOneWriteOpResult<any>> {
+function insert (data: any): Promise<InsertOneWriteOpResult<any>> {
   return db.collection('users').insertOne(data)
 }
 
-export function update (query: object, data: any): Promise<UpdateWriteOpResult> {
+function update (query: object, data: any): Promise<UpdateWriteOpResult> {
   return db.collection('users').updateOne(query, data)
+}
+
+export default {
+  parseId,
+  init,
+  get,
+  insert,
+  update
 }
