+++ server/src/services/spot-data.ts
